Start factoring out legacy_xml.rs
